Z4XV8R8 un-merge command-utils: persist root and assignable contributors are separate questions
