+++ source/lib/command-line/commands/persist-root.ts
@@ -1,0 +1,9 @@
+import {isFail, succeeded} from '../../model/result-types.js';
+import {getPersistRoot} from '../../storage/paths.js';
+
+export const getPersistRootValue = async () => {
+	const persistRootResult = await getPersistRoot();
+	if (isFail(persistRootResult)) return persistRootResult;
+
+	return succeeded('Resolved persist root', persistRootResult.value);
+};
