+++ source/lib/command-line/commands/unassign-user.cmd.ts
@@ -6,10 +6,8 @@ import {failed, isFail} from '../../model/result-types.js';
 import {findAncestor} from '../../repository/node-repo.js';
 import {getCmdState} from '../../state/cmd.state.js';
 import {getState} from '../../state/state.js';
-import {
-	getAssignableContributors,
-	getPersistRootValue,
-} from './command-utils.js';
+import {getAssignableContributors} from './assignable-contributors.js';
+import {getPersistRootValue} from './persist-root.js';
 
 export const unassignUserCommand = async () => {
 	const userRes = resolveActorId();
